ODROID: bios: updating SPI flash memory from Linux user space
This patch is to update SPI flash memory image from Linux userspace rather than
U-boot itslef previously implemented. The flash memory image file 'spiboot.img'
must be copied to '/dev/mmcblk0p1' which is the first partition of the SD card,
not eMMC.

The eMMC is not the storage to scan the image file at all, since eMMC and SPI
flash memory share the hardware bus and both cannot be accessible.

Change-Id: If297bb6896959cb2c2ccbb30ad01da3202c6e218
Signed-off-by: Dongjin Kim <tobetter@gmail.com>